[autodiscovery/integration/config] Add json tag to ServiceID (#18364)
* [autodiscovery/integration/config] Add json tag to ServiceID

The missing json tag breaks the cluster checks rebalancing algorithm for
service checks.

ServiceID is used to calculate the digest of a check. In service checks it
follows this format: `kube_service://<namespace>/<name>`. The problem is that,
because of the missing json tag, this field is not propagated to the check
runners, so they calculate a different digest for the same check.

As a consequence, this breaks the cluster checks rebalancing algorithm, because
it doesn't detect service checks as cluster checks, so they're never considered
in rebalancing decisions (Reference:
https://github.com/DataDog/datadog-agent/blob/29fd4466c00da9f6ef1768c1337aa93cb15229cd/pkg/clusteragent/clusterchecks/dispatcher_nodes.go#L191)